Refactor compiled branch of runPanel to separate file.
