+++ gui/src/app/RunPanel/RunPanel.tsx
@@ -1,22 +1,17 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import Box from "@mui/material/Box";
-import LinearProgress, {
-  LinearProgressProps,
-} from "@mui/material/LinearProgress";
-import Typography from "@mui/material/Typography";
 import { FunctionComponent, useCallback, useContext } from "react";
 
+import Button from "@mui/material/Button";
+import { CompileContext } from "@SpCompileContext/CompileContext";
+import { ProjectContext } from "@SpCore/ProjectContextProvider";
 import {
   ProjectDataModel,
   SamplingOpts,
   modelHasUnsavedChanges,
 } from "@SpCore/ProjectDataModel";
-import { Progress } from "@SpStanSampler/StanModelWorker";
 import StanSampler from "@SpStanSampler/StanSampler";
 import { StanRun } from "@SpStanSampler/useStanSampler";
-import Button from "@mui/material/Button";
-import { CompileContext } from "@SpCompileContext/CompileContext";
-import { ProjectContext } from "@SpCore/ProjectContextProvider";
+import CompiledRunPanel from "./CompiledRunPanel";
 
 type RunPanelProps = {
   sampler?: StanSampler;
@@ -57,48 +52,14 @@ const RunPanel: FunctionComponent<RunPanelProps> = ({
   let content;
   if (compileStatus === "compiled") {
     content = (
-      <div>
-        <Button
-          variant="contained"
-          color="success"
-          onClick={handleRun}
-          disabled={runStatus === "sampling" || runStatus === "loading"}
-        >
-          run sampling
-        </Button>
-        &nbsp;
-        {runStatus === "sampling" && (
-          <Button
-            color="error"
-            variant="outlined"
-            onClick={cancelRun}
-            disabled={runStatus !== "sampling"}
-          >
-            cancel
-          </Button>
-        )}
-        <hr />
-        {runStatus === "loading" && <div>Loading compiled Stan model...</div>}
-        {runStatus === "sampling" && (
-          <div>
-            Sampling
-            <SamplingProgressComponent
-              report={progress}
-              numChains={samplingOpts.num_chains}
-            />
-          </div>
-        )}
-        {runStatus === "completed" && <div>done sampling</div>}
-        {runStatus === "failed" && (
-          <div>
-            Sampling failed!
-            <pre className="SamplerError">{errorMessage}</pre>
-            <span className="details">
-              (see browser console for more details)
-            </span>
-          </div>
-        )}
-      </div>
+      <CompiledRunPanel
+        handleRun={handleRun}
+        cancelRun={cancelRun}
+        runStatus={runStatus}
+        progress={progress}
+        samplingOpts={samplingOpts}
+        errorMessage={errorMessage}
+      />
     );
   } else if (["preparing", "compiling"].includes(compileStatus)) {
     content = <div>{compileMessage}</div>;
@@ -131,59 +92,6 @@ const isCompileModelDisabled = (
   if (modelHasUnsavedChanges(projectData)) return true;
   if (!validSyntax) return true;
   return false;
-};
-
-type SamplingProgressComponentProps = {
-  report: Progress | undefined;
-  numChains: number;
-};
-
-const SamplingProgressComponent: FunctionComponent<
-  SamplingProgressComponentProps
-> = ({ report, numChains }) => {
-  if (!report) return <span />;
-  const progress =
-    ((report.iteration + (report.chain - 1) * report.totalIterations) /
-      (report.totalIterations * numChains)) *
-    100;
-  return (
-    <>
-      <div className="SamplingProgress">
-        <LinearProgressWithLabel
-          sx={{
-            height: 10,
-            // https://stackoverflow.com/a/73009519
-            "& .MuiLinearProgress-bar": {
-              transition: "none",
-            },
-          }}
-          value={progress}
-        />
-      </div>
-      <div>
-        Chain {report.chain} Iteration: {report.iteration} /{" "}
-        {report.totalIterations} ({report.warmup ? "Warmup" : "Sampling"})
-      </div>
-    </>
-  );
-};
-
-// from https://mui.com/material-ui/react-progress/#linear-with-label
-const LinearProgressWithLabel = (
-  props: LinearProgressProps & { value: number },
-) => {
-  return (
-    <Box sx={{ display: "flex", alignItems: "center" }}>
-      <Box sx={{ width: "100%", mr: 1 }}>
-        <LinearProgress variant="determinate" {...props} />
-      </Box>
-      <Box sx={{ minWidth: 35 }}>
-        <Typography variant="body2" color="text.secondary">{`${Math.round(
-          props.value,
-        )}%`}</Typography>
-      </Box>
-    </Box>
-  );
 };
 
 export default RunPanel;
